+++ client/src/apis/apiUtils.ts
@@ -1,15 +1,6 @@
 import 'dotenv/config';
-import { getOptions, postOptions, patchOptions, deleteOptions, BodyType } from './options';
-const baseUrl = process.env.REACT_APP_SERVER_URL ?? '';
 
-export interface HTTPResponse<T> {
-  isOk: boolean;
-  errorData?: {
-    message: string;
-    statusCode: number;
-  };
-  data?: T;
-}
+const baseUrl = process.env.REACT_APP_SERVER_URL ?? '';
 
 export function getUrl(url: string): string {
   return baseUrl + url;
@@ -19,56 +10,6 @@ export function queryObjToString<T>(queryObj: T): string {
   return Object.entries(queryObj)
     .map((e) => e.join('='))
     .join('&');
-}
-
-export async function fetcher<T>(url: string, options: RequestInit): Promise<HTTPResponse<T>> {
-  try {
-    const response = await fetch(url, options);
-    const { statusCode, data, message } = await response.json();
-    if (response.ok) {
-      const res = { isOk: true, data };
-      return res;
-    }
-    return {
-      isOk: false,
-      errorData: {
-        message,
-        statusCode,
-      },
-    };
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  } catch (e: any) {
-    return {
-      isOk: false,
-      errorData: {
-        message: e.message,
-        statusCode: e.statusCode,
-      },
-    };
-  }
-}
-
-export async function fetchGet<T>(url: string, query?: string): Promise<HTTPResponse<T>> {
-  const requestUrl = getUrl(`${url}?${query}`);
-  const response = await fetcher<T>(requestUrl, getOptions());
-  return response;
-}
-
-export async function fetchPost<T>(url: string, body: BodyType = {}): Promise<HTTPResponse<T>> {
-  const requestUrl = getUrl(url);
-  const response = await fetcher<T>(requestUrl, postOptions(body));
-  return response;
-}
-
-export async function fetchPatch<T>(url: string, body: BodyType = {}): Promise<HTTPResponse<T>> {
-  const requestUrl = getUrl(url);
-  const response = await fetcher<T>(requestUrl, patchOptions(body));
-  return response;
-}
-
-export function fetchDelete(url: string): void {
-  const requestUrl = getUrl(url);
-  fetcher(requestUrl, deleteOptions());
 }
 
 type TypeRoom = '타닥타닥' | '캠프파이어';
@@ -88,10 +29,4 @@ export function getRoomQueryObj(type: TypeRoom, search: string, page: number): Q
     page,
   };
   return queryObj;
-}
-
-export async function fetchDeleteImage<T>(url: string): Promise<HTTPResponse<T>> {
-  const requestUrl = getUrl(url);
-  const response = await fetcher<T>(requestUrl, deleteOptions());
-  return response;
 }
