+++ client/src/types/httpResponse.ts
@@ -1,0 +1,8 @@
+export default interface HTTPResponse<T> {
+  isOk: boolean;
+  errorData?: {
+    message: string;
+    statusCode: number;
+  };
+  data?: T;
+}
