+++ client/src/types/index.ts
@@ -2,5 +2,6 @@ import { DevFieldType } from './devField';
 import RoomInfoType from './roomInfo';
 import UserInfoType from './userInfo';
 import TabStateType from './tabState';
+import HTTPResponse from './httpResponse';
 
-export type { DevFieldType, RoomInfoType, UserInfoType, TabStateType };
+export type { DevFieldType, RoomInfoType, UserInfoType, TabStateType, HTTPResponse };
